requirements.txt
absl-py==1.4.0
asttokens==2.2.1
backcall==0.2.0
blessed==1.20.0
boto3==1.26.145
botocore==1.29.145
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cmake==3.25.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
filelock==3.10.3
flake8==6.0.0
fonttools==4.39.3
ftfy==6.1.1
google-auth==2.16.3
google-auth-oauthlib==0.4.6
gpustat==1.0.0
grpcio==1.51.3
h5py==3.6.0
huggingface-hub==0.13.3
idna==3.4
imageio==2.28.1
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy_loader==0.2
lit==15.0.7
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
modules==1.0.0
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==3.0
numpy==1.21.4
nvidia-ml-py==11.495.46
oauthlib==3.2.2
packaging==23.0
pandas==1.3.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-pretrained-bert==0.6.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
sacremoses==0.0.53
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.10.3
tornado==6.2
tqdm==4.62.3
traitlets==5.9.0
transformers==4.12.5
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
yacs==0.1.8
yapf==0.32.0
zipp==3.15.0
